server/requirements.txt
flask==2.0.1
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
werkzeug==2.0.1
gunicorn==20.1.0
